+++ CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.13](https://github.com/helium/modular-governance/compare/v0.0.12...v0.0.13) (2024-09-06)
+
+**Note:** Version bump only for package modular-governance
+
+
+
+
+
+## [0.0.12](https://github.com/helium/modular-governance/compare/v0.0.10...v0.0.12) (2024-08-12)
+
+
+
+## [0.0.8](https://github.com/helium/modular-governance/compare/v0.0.7...v0.0.8) (2023-11-10)
+
+
+
+## [0.0.7](https://github.com/helium/modular-governance/compare/v0.0.6...v0.0.7) (2023-11-07)
+
+
+
+## [0.0.6](https://github.com/helium/modular-governance/compare/v0.0.4...v0.0.6) (2023-11-06)
+
+**Note:** Version bump only for package modular-governance
+
+
+
+
+
 ## [0.0.11](https://github.com/helium/modular-governance/compare/v0.0.10...v0.0.11) (2024-06-17)
 
 **Note:** Version bump only for package modular-governance
@@ -14,8 +42,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## [0.0.10](https://github.com/helium/modular-governance/compare/v0.0.9...v0.0.10) (2024-03-28)
 
 **Note:** Version bump only for package modular-governance
-
-